backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.12
email-validator>=2.0.0
